% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_functions.R
\name{get_scaled_covars}
\alias{get_scaled_covars}
\title{Scale a covariate matrix}
\usage{
get_scaled_covars(X)
}
\arguments{
\item{X}{The covariate matrix}
}
\value{
A list with 2 elements: the scaled matrix and a dataframe with the
  mean and standard deviation used to scale the matrix.
}
\description{
Scale a covariate matrix
}

