% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_plots.R
\name{make_coef_plot}
\alias{make_coef_plot}
\title{Generate coefficient path plot.}
\usage{
make_coef_plot(coef_df, best_lambda, style = "base")
}
\arguments{
\item{coef_df}{A data frame with 3 columns: \code{covariate}, \code{lambda},
\code{value}.}

\item{best_lambda}{A scalar for the lambda value at which the cv score is
lowest. For plotting a vertical line for the selected coefficient values.}

\item{style}{A character scalar. Possible values are \code{c("base", "clean")}.
\code{"base"} returns a ggplot object using the base (gray) theme. \code{"clean"}
returns a ggplot object using \code{theme_few}. Default is \code{"base"}.}
}
\value{
A ggplot object.
}
\description{
Generate coefficient path plot.
}
