% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_plots.R
\name{make_cvscore_plot}
\alias{make_cvscore_plot}
\title{Generate cross-validation score path plot.}
\usage{
make_cvscore_plot(cvscore_df, score_name, best_lambda, style = "base")
}
\arguments{
\item{cvscore_df}{A data frame with 2 columns: \code{cvscore}, \code{lambda}.}

\item{score_name}{A string identifying the type of score used, e.g.,
\code{"MSE"} for mean square error.}

\item{best_lambda}{A scalar for the lambda value at which the cv score is
lowest. For plotting a vertical line for the selected coefficient values.}

\item{style}{A character scalar. Possible values are \code{c("base", "clean")}.
\code{"base"} returns a ggplot object using the base (gray) theme. \code{"clean"}
returns a ggplot object using \code{theme_few}. Default is \code{"base"}.}
}
\value{
A ggplot object.
}
\description{
Generate cross-validation score path plot.
}
