% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_functions.R
\name{scale_oos_covars}
\alias{scale_oos_covars}
\title{Scale an out-of-sample covariate vector}
\usage{
scale_oos_covars(X, params)
}
\arguments{
\item{params}{A k-by-2 matrix containing the means of the in-sample covariate
matrix in the first column and the sds of the in-sample covariate matrix
in the second column.}

\item{x}{A covariate matrix.}
}
\value{
The scaled covariate vector.
}
\description{
Scale an out-of-sample covariate vector
}

