% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_functions.R
\name{waic}
\alias{waic}
\title{Calculate likelihood-based metrics for WAIC and LPD}
\usage{
waic(stanfit)
}
\arguments{
\item{stanfit}{A formal class stanfit}
}
\value{
A list of WAIC and LPD metrics
}
\description{
Calculate likelihood-based metrics for WAIC and LPD
}

